chore: rename files of like to postLike
